:art: SEA-22: isolate menu button into its own component for nav
